Implement pause/resume
This change introduces the ability to pause and resume playback from
the calling page. This turns out to be a bit more roundabout than
expected. The web audio API apparently has no support for pause/resume.
Thus, it has to be implemented using a flag to indicate a paused state,
signaling to the audio generation callback to stuff 0 values into the
output buffer while the visualizer stops requesting new animation
frames.

It works but could still use optimization, especially in the silent
audio generation. This should track whether the entire buffer has been
filled with 0s already and skip filling again.

So far, this works equally well on Chrome, Firefox, and Safari.