Add an API function for hiding the viz
While the recent refactoring gave greater control to the client app to
set the visualization (including disabling it manually), something else
is needed to account for when the player tab is hidden. This is
different than disabling the viz because some layer of the stack needs
to remember the viz state for when the tab is revealed again.